% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\docType{data}
\name{aware}
\alias{aware}
\title{English Adaptation of the WHO AWaRe Index}
\format{A data frame with 208 antimicrobial agents, referenced 
against the Anatomical Therapeutic Chemical (ATC) 
classification \insertCite{WHO-ATC2020}{Ramses} and the 
NHS Dictionary of Medicines (dm+d) \insertCite{DMD2015}{Ramses}.

It contains 6 columns:
\describe{
\item{\code{ATC_name}}{ATC name of the antibiotic}
\item{\code{ATC_code}}{ATC code of the antibiotic}
\item{\code{ATC_route}}{route of administration as defined in
the ATC ("O" = oral; "P" = parenteral; "R" = rectal;
"V" = vaginal)}
\item{\code{AWaRe_England}}{AWaRe index assigned by Public Health
England in 2018 ("Access", "Watch", "Reserve")}
\item{\code{VTM_code}}{antibiotic Virtual Therapeutic Moiety 
concept code in the SNOMED-CT UK Drug Extension (see \insertCite{DMD2015}{Ramses})}
\item{\code{VTM_name}}{antibiotic Virtual Therapeutic Moiety 
concept label in the SNOMED-CT UK Drug Extension (see \insertCite{DMD2015}{Ramses})}
}}
\source{
Adapted from \insertRef{PHE-aware2019}{Ramses} with Virtual  
Therapeutic Moiety equivalents from the NHS Dictionary of Medicines (dm+d)
\insertCite{DMD2015}{Ramses}.
}
\usage{
aware
}
\description{
This `AWaRe Index' is a classification by the World Health
Organisation of key antibiotics into three categories
\insertCite{Sharland2018}{Ramses} to:
\itemize{
       \item improve access (Access)
       \item monitor important antibiotics (Watch) and 
       \item preserve effectiveness of ‘last resort’ antibiotics (Reserve).
}
It was adapted by Public Health England for antibiotic stewardship
and quality improvement in English hospitals 
\insertCite{PHE-aware2019}{Ramses}
}
\references{
{\insertAllCited{}}
}
\keyword{datasets}
