% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ramses-deprecated.R
\name{bridge_spell_therapy_overlap}
\alias{bridge_spell_therapy_overlap}
\title{#' Create database bridge tables}
\usage{
bridge_spell_therapy_overlap(conn, overwrite = FALSE)
}
\arguments{
\item{conn}{a database connection}

\item{overwrite}{if \code{TRUE}, will overwrite any existing
database table. The default is \code{FALSE}}
}
\description{
This function is now deprecated. Please use
 \code{\link{bridge_encounter_therapy_overlap}()}.
}
\seealso{
\code{\link{bridge_tables}}, \code{\link{Ramses-deprecated}}
}
