% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{build_ramses_schema}
\alias{build_ramses_schema}
\title{Build the RAMSES schema}
\usage{
build_ramses_schema(conn)
}
\arguments{
\item{conn}{a database connection}
}
\description{
The RAMSES schema is the set of tables required to import health records
for curation and analysis
}
