% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_features.R
\name{clinical_feature_last}
\alias{clinical_feature_last}
\alias{clinical_feature_last,TherapyEpisode-method}
\title{Therapy table feature: most recent observation}
\usage{
clinical_feature_last(x, investigation_codes, hours)

\S4method{clinical_feature_last}{TherapyEpisode}(x, investigation_codes, hours)
}
\arguments{
\item{x}{a \link{TherapyEpisode} object}

\item{investigation_codes}{a character vector of clinical investigation codes
matching the \code{observation_code} field in the \code{inpatient_investigation}
table (see \code{\link{validate_investigations}()}}

\item{hours}{the maximum number of hours the observation should date back from
\code{t_start}, the starting time of every row in the therapy table}
}
\value{

}
\description{
Add a clinical feature to a therapy episode table indicating 
the most recent observation before table time \code{t}
}
\details{
NOTE: only numeric investigations marked with status \code{"final"}, 
\code{"preliminary"}, \code{"corrected"}, or \code{"amended"} will be sourced.
}
