% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_features.R
\name{clinical_feature_ols_trend}
\alias{clinical_feature_ols_trend}
\alias{clinical_feature_ols_trend,TherapyEpisode-method}
\title{Therapy table feature: temporal trend of clinical observations}
\usage{
clinical_feature_ols_trend(x, observation_code, hours)

\S4method{clinical_feature_ols_trend}{TherapyEpisode}(x, observation_code, hours)
}
\arguments{
\item{x}{a \code{\link{TherapyEpisode}} object}

\item{observation_code}{a character vector of clinical investigation codes
matching the \code{observation_code} field in the \code{inpatient_investigation}
table (see \code{\link{validate_investigations}()}}

\item{hours}{the maximum number of hours the observations should date back from
\code{t_start}, the starting time of every row in the therapy table}
}
\value{
an object of class \code{\link{TherapyEpisode}}
}
\description{
Add a clinical feature to a therapy episode table corresponding
to the Ordinary Least Squares (OLS) intercept and slope of clinical
observations made until table time \code{t}
}
\details{
NOTE: only numeric investigations marked with status \code{"final"}, 
\code{"preliminary"}, \code{"corrected"}, or \code{"amended"} will be sourced.
}
