% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{compute}
\alias{compute}
\alias{compute,TherapyEpisode-method}
\alias{compute,Encounter-method}
\alias{compute,RamsesObject-method}
\title{Compute a Ramses Object}
\usage{
compute(x, ...)

\S4method{compute}{TherapyEpisode}(x)

\S4method{compute}{Encounter}(x)

\S4method{compute}{RamsesObject}(x)
}
\arguments{
\item{x}{an object of class \code{RamsesObject}}

\item{...}{arguments passed on to methods (compatibility with \code{dplyr}'s
S3 generic)}
}
\value{
an object of class \code{RamsesObject}
}
\description{
Compute a Ramses Object
}
\seealso{
\code{\link[dplyr]{compute}()}
}
