% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{compute_DDDs}
\alias{compute_DDDs}
\title{Compute Defined Daily Doses (DDDs)}
\usage{
compute_DDDs(ab, administration, dose, unit)
}
\arguments{
\item{ab}{a character vector coercible to an antibiotic code using 
\code{\link[AMR]{as.ab}()}}

\item{administration}{a character vector indicating the route of 
administration (see Details)}

\item{dose}{a numeric vector indicating the total dose for the basis 
of strength of the ATC DDD: in the case of prescriptions, this should
be the total dose given in one day}

\item{unit}{a character vector coercible to a `units` object with 
\code{\link[units]{as.units}()}.}
}
\value{
a numeric vector containing the DDDs
}
\description{
Compute Defined Daily Doses (DDDs) for a drug prescription 
or administration record using the World Health Organisation Anatomical 
Therapeutic Chemical (ATC) classification \insertCite{WHO-ATC2020}{Ramses}.
}
\details{
This function queries the WHO ATC website for the most 
up-to-date reference values of the DDDs.
  
The `administration` parameter must be one of the following values:
 \itemize{
     \item \code{'Implant'} for implants
     \item \code{'Inhal'} for inhalation (eg: nebuliser or inhalers)
     \item \code{'Instill'} for instillation (eg: topical drops)
     \item \code{'N'} for nasal administration
     \item \code{'O'} for oral administration or administration via 
     percutaneous endoscopic gastrostomy tube 
     \item \code{'P'} for parenteral administration (eg intravenous 
     bolus/infusion injections, intramuscular injections)
     \item \code{'R'} for rectal administration
     \item \code{'SL'} for sublingual/buccal/oromucosal administration
     \item \code{'TD'} for transdermal administration (eg: patch)
     \item \code{'V'} for vaginal administration
 }
}
\examples{
# Three tablets of amoxicillin 500mg
compute_DDDs("Amoxicillin", "O", 1.5, "g")

# Three tablets of co-amoxiclav 625mg (containing 500mg amoxicillin each)
compute_DDDs("Coamoxiclav", "O", 1.5, "g") 
}
\references{
\insertAllCited{}
}
