% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{connect_db_local}
\alias{connect_db_local}
\title{Create a local database}
\usage{
connect_db_local(file)
}
\arguments{
\item{file}{A file path to an existing or new database file with an .sqlite extension.}
}
\value{
An object of class SQLiteConnection.
}
\description{
Create a local database in memory or on disk using RSQLite. This is 
the easiest method to work on a small scale, with a limited number of users.
}
\details{
This function creates a database on disk at the desired path. The database and
its content will persist on disconnection.
}
\section{Warning}{
     
This method does not provide any encryption or password protection. For work with 
real patient data, only use suitably encrypted databases.
}

\examples{
    # Create database and load data
    con <- connect_db_local("ramses-db.sqlite")
    
    dplyr::copy_to(dest = con, df = aware, name = "LU_AWaRe", 
                   overwrite = FALSE, temporary = FALSE)      
    
    # Close connection to database
    DBI::dbDisconnect(con)
    
    # Connect to the database again and show data
    con <- connect_db_local("ramses-db.sqlite")
    dplyr::tbl(con, "LU_AWaRe")
    DBI::dbDisconnect(con)
}
\seealso{
The dbplyr website provides excellent guidance on how to connect to databases: 
\url{https://db.rstudio.com/getting-started/connect-to-database}
}
