% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{connect_local_database}
\alias{connect_local_database}
\title{Create or connect to a local SQLite database}
\usage{
connect_local_database(file)
}
\arguments{
\item{file}{A file path to an existing or new database file with an .sqlite extension.}
}
\value{
An object of class SQLiteConnection.
}
\description{
Create a local database in memory or on disk using \code{\link[RSQLite]{SQLite}()}. 
This is the ideal method to experiment on a small scale or for testing purposes.
}
\details{
This function creates a database on disk at the desired path. The database and
its content will persist on disconnection.
}
\section{Warning}{
     
This method does not provide any encryption or password protection. You should only use this
method with mock data unless you operate within a secure data enclave.
}

\examples{
    # Create database and load data
    con <- connect_local_database("ramses-db.sqlite")
    
    dplyr::copy_to(dest = con, df = reference_aware, name = "reference_aware", 
                   overwrite = FALSE, temporary = FALSE)      
    
    # Close connection to database
    DBI::dbDisconnect(con)
    
    # Connect to the database again and show data
    con <- connect_local_database("ramses-db.sqlite")
    dplyr::tbl(con, "reference_aware")
    DBI::dbDisconnect(con)
}
\seealso{
The dbplyr website provides excellent guidance on how to connect to databases: 
\url{https://db.rstudio.com/getting-started/connect-to-database}
}
