% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge_and_metrics.R
\name{create_reporting_inpatient}
\alias{create_reporting_inpatient}
\title{Build a reporting hypercube for inpatient activity}
\usage{
create_reporting_inpatient(conn)
}
\arguments{
\item{conn}{a database connection}
}
\value{
a object of \code{`tbl_sql`} connected to the \code{reporting_inpatient_length_of_stay} table
}
\description{
Create or overwrite table \code{`hypercube_inpatient`} which reports 
on prescribing during inpatient encounters.
}
\details{
This function may be used after loading records into the Ramses database
to facilitate reporting on inpatient activity.

DIMENSIONs
The following dimensions are built, if available from the Ramses database
}
