% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge_and_metrics.R
\name{create_reporting_med_prescribing}
\alias{create_reporting_med_prescribing}
\title{Create prescribing reporting table}
\usage{
create_reporting_med_prescribing(conn)
}
\arguments{
\item{conn}{a database connection}
}
\value{
a object of \code{`tbl_sql`} connected to the \code{metrics_prescribing} table
}
\description{
Create prescribing reporting table
}
\details{
#' METRICS

\itemize{
    \item{DOT_prescribed: }
    \item{DOT_administered: } IF AVAILABLE
    \item{DDD_prescribed: } IF AVAILABLE
    \item{DDD_administered: } IF AVAILABLE
}
}
