% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{create_therapy_episodes}
\alias{create_therapy_episodes}
\alias{create_therapy_episodes.SQLiteConnection}
\alias{create_therapy_episodes.PqConnection}
\title{Re-create therapy episodes and therapy combinations}
\usage{
create_therapy_episodes(
  conn,
  transitive_closure_controls = transitive_closure_control(),
  silent = TRUE
)

\method{create_therapy_episodes}{SQLiteConnection}(
  conn,
  transitive_closure_controls = transitive_closure_control(),
  silent = TRUE
)

\method{create_therapy_episodes}{PqConnection}(
  conn,
  transitive_closure_controls = transitive_closure_control(),
  silent = TRUE
)
}
\arguments{
\item{conn}{a connection to a database containing a
\code{drug_prescriptions} table previously created with 
\code{\link{load_medications}()})}

\item{transitive_closure_controls}{parameters controlling (see 
\code{\link{transitive_closure_control}()})}

\item{silent}{a boolean indicating whether the function should be
executed without progress bar. Default is \code{TRUE}.}
}
\description{
This function should be used to recreate therapy episodes
and therapy combinations, for example in order to try new 
\code{transitive_closure_controls} parameters. Note that 
\code{\link[Ramses]{load_medications}()} creates 
episodes and combinations by default.
}
\details{
This function performs the following tasks:
\enumerate{
   \item verify that a valid \code{drug_prescriptions} table exists
   \item delete any existing \code{drug_prescription_edges} and
   \code{drug_therapy_episodes} table
   \item recreate the \code{drug_prescription_edges} table
   \item repopulate the \code{therapy_id} and \code{combination_id}
   fields in \code{drug_prescriptions} table
   \item recreate the \code{drug_therapy_episodes} table
}
}
