% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{Encounter-class}
\alias{Encounter-class}
\alias{Encounter}
\title{An S4 class to represent inpatient encounters}
\usage{
Encounter(conn, id)
}
\arguments{
\item{conn}{a database connection}

\item{id}{an encounter identifier}
}
\description{
An S4 class to represent inpatient encounters
}
\section{Slots}{

\describe{
\item{\code{id}}{an encounter identifier}

\item{\code{conn}}{a database connection}

\item{\code{record}}{a \code{tbl_sql} for the corresponding database record}

\item{\code{longitudinal_table}}{a \code{tbl_sql} for the longitudinal encounter table}
}}

