% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_ATC_name}
\alias{get_ATC_name}
\title{Look up ATC name on the WHO website}
\usage{
get_ATC_name(x, silent = FALSE)
}
\arguments{
\item{x}{a character vector of ATC codes}

\item{silent}{if \code{TRUE}, the progress bar will be hidden. 
The default is \code{FALSE}.}
}
\value{
a character vector of ATC names (lowercase).
}
\description{
Look up drug name from an ATC code using the WHO DDD website.
Requires an internet connection.
}
\examples{
get_ATC_name("J01XA01")
}
\references{
Adapted from \code{\link[AMR]{atc_online_property}()} with thanks to 
the \link{AMR} package authors.
}
