% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{get_therapy_table}
\alias{get_therapy_table}
\alias{get_therapy_table,TherapyEpisode-method}
\title{Get the therapy table}
\usage{
get_therapy_table(object, collect = FALSE)

\S4method{get_therapy_table}{TherapyEpisode}(object, collect = FALSE)
}
\arguments{
\item{object}{an object of class \code{TherapyEpisode}}

\item{collect}{whether to collect the remote \code{tbl_sql} and return a local 
\code{tbl_df}. The default is \code{FALSE}}
}
\value{
an object of class \code{tbl}
}
\description{
Get the therapy table
}
