% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_warehouse_status}
\alias{get_warehouse_status}
\title{Get warehouse status}
\usage{
get_warehouse_status(conn)
}
\arguments{
\item{conn}{a database connection}
}
\value{
\describe{
    \item{diagnoses}{a flag indicating whether }
  }
}
\description{
Inspect the status of tables making up the warehouse
}
