% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd.R
\name{import_icd}
\alias{import_icd}
\title{International Classification of Diseases}
\usage{
import_icd(archive, version)
}
\arguments{
\item{archive}{path to the ZIP archive of the ICD-10 release 
(5th Edition is recommended)}

\item{version}{a character string labelling the edition 
(e.g. "GB 5th Edition")}
}
\value{
A data frame ready for importing into the database.
}
\description{
The World Health Organisation (WHO) International 
Statistical Classification of Diseases and Related Health 
Problems Tenth Revision is available under restrictive licence terms.

The UK implementation of ICD-10 is available from the NHS TRUD website
\url{https://isd.digital.nhs.uk/trud3/user/guest/group/0/pack/28} 
at no cost to UK users subject to a licence agreement.
 
The US implementation of ICD-10-CM is available from 
\url{https://www.cdc.gov/nchs/icd/icd10cm.htm}
}
\examples{
\dontrun{
import_icd("icd_df_10.5.0_20151102000001.zip", "GB 5th Edition")
}
}
