% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_records.R
\docType{data}
\name{inpatient_diagnoses}
\alias{inpatient_diagnoses}
\title{Simulated inpatient clinical diagnoses}
\format{
A data frame of records of clinical diagnoses.
}
\source{
Simulated data
}
\usage{
inpatient_diagnoses
}
\description{
Dataset of simulated infection clinical diagnoses by 
finished consultant episodes listed in 
\code{\link[Ramses]{inpatient_episodes}},
detailing all episode start and end times, episode number, and up to
5 clinical diagnoses coded according to the International 
Classification of Diseases, Tenth Revision (ICD-10). 

This data was simulated to reproduce a set of combinations of infection 
diagnoses in an English NHS hospital. For each episode, a maximum of 5
diagnoses are simulated. Non-infection diagnoses are not simulated, an
NA value is recorded instead.
}
\keyword{datasets}
