% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_records.R
\docType{data}
\name{inpatient_investigations}
\alias{inpatient_investigations}
\title{Simulated clinical investigation results}
\format{
A data frame of records of observations and investigations 
passing \code{\link[Ramses]{validate_investigations}()}.
}
\source{
Simulated data
}
\usage{
inpatient_investigations
}
\description{
A set of records of observations (for example: respiration rate,
blood pressure) and investigations (for example: blood cell counts)
performed during inpatient care episodes.
}
\keyword{datasets}
