% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_records.R
\docType{data}
\name{inpatient_microbiology}
\alias{inpatient_microbiology}
\title{Simulated microbial culture and susceptibility results}
\format{
A data frame
}
\source{
Simulated data
}
\usage{
inpatient_microbiology
}
\description{
This simulated activity data consists of records of microbial 
samples sent for culture and susceptibility testing for a single inpatient.

This data was simulated to reproduce relevant characteristics of English 
NHS Admitted Patient Care datasets. More information can be found on the 
\href{https://datadictionary.nhs.uk/data_sets/cds_v6-2/cds_v6-2_type_130_-_admitted_patient_care_-_finished_general_episode_cds.html#dataset_cds_v6-2_type_130_-_admitted_patient_care_-_finished_general_episode_cds}{NHS data dictionary website}.
}
\seealso{
\link[Ramses]{validate_microbiology}()
}
\keyword{datasets}
