% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{load_inpatient_episodes}
\alias{load_inpatient_episodes}
\title{Load episodes of care records into the warehouse}
\usage{
load_inpatient_episodes(
  conn,
  patients_data,
  episodes_data,
  wards_data = NULL,
  overwrite = TRUE
)
}
\arguments{
\item{conn}{a database connection}

\item{patients_data}{a data frame validated with
\code{\link{validate_inpatient_episodes}()}}

\item{episodes_data}{a data frame validated with 
\code{\link{validate_inpatient_episodes}()}}

\item{wards_data}{a data frame validated with
\code{\link{validate_inpatient_episodes}()}#'}

\item{overwrite}{if \code{TRUE} (the default), will overwrite any existing
\code{inpatient_episodes} database table}
}
\description{
Load episodes of care records into the warehouse
}
