% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{load_inpatient_investigations}
\alias{load_inpatient_investigations}
\title{Load observations & investigations into the warehouse}
\usage{
load_inpatient_investigations(conn, investigations_data, overwrite = TRUE)
}
\arguments{
\item{conn}{a database connection}

\item{investigations_data}{a data frame validated with 
\code{\link{validate_investigations}()}}

\item{overwrite}{if \code{TRUE} (the default), will overwrite any existing
\code{inpatient_investigations} database table}
}
\description{
Load observations & investigations into the warehouse
}
\seealso{
\code{\link{validate_investigations}()}
}
