% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{load_inpatient_microbiology}
\alias{load_inpatient_microbiology}
\title{Load observations & investigations into the warehouse}
\usage{
load_inpatient_microbiology(
  conn,
  specimens,
  isolates,
  susceptibilities,
  overwrite = TRUE
)
}
\arguments{
\item{conn}{a database connection}

\item{specimens}{a data frame with one row per specimen sent
to laboratory and
validated with \code{\link{validate_microbiology}()}}

\item{isolates}{a data frame with one row per microorganism
isolated from the laboratory specimen and
validated with \code{\link{validate_microbiology}()}}

\item{susceptibilities}{a data frame with one row per susceptibility 
validated with \code{\link{validate_microbiology}()}}

\item{overwrite}{if `TRUE` (the default), will overwrite any existing
`inpatient_investigations` database table}
}
\value{
`TRUE` if the function ran successfully, otherwise object of class
"try-error" containing error messages trigger during warehouse data loading.
}
\description{
Load observations & investigations into the warehouse
}
\seealso{
\code{\link{validate_investigations}()},
}
