% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{load_medications}
\alias{load_medications}
\alias{load_medications.SQLiteConnection}
\alias{load_medications.PqConnection}
\title{Load drug prescription records into the warehouse}
\usage{
load_medications(
  conn,
  prescriptions,
  administrations = NULL,
  overwrite = FALSE,
  transitive_closure_controls = transitive_closure_control(),
  silent = FALSE
)

\method{load_medications}{SQLiteConnection}(
  conn,
  prescriptions,
  administrations = NULL,
  overwrite = FALSE,
  transitive_closure_controls = transitive_closure_control(),
  silent = TRUE
)

\method{load_medications}{PqConnection}(
  conn,
  prescriptions,
  administrations = NULL,
  overwrite = FALSE,
  transitive_closure_controls = transitive_closure_control(),
  silent = TRUE
)
}
\arguments{
\item{conn}{a database connection}

\item{prescriptions}{a data frame of prescriptions passing the 
\code{\link{validate_prescriptions}()} function. Note: if variables 
\code{therapy_id} or \code{combination_id} are provided, they will be preserved
as they are instead of being populated by Ramses}

\item{administrations}{a data frame of drug administrations passing the 
\code{\link{validate_administrations}()} function}

\item{overwrite}{if \code{TRUE} (the default), will overwrite existing 
medication tables}

\item{transitive_closure_controls}{parameters controlling (see 
\code{\link{transitive_closure_control}()})}

\item{silent}{if \code{TRUE}, the progress bar will be hidden. The default is 
\code{FALSE}.}
}
\description{
Load drug prescription records into the warehouse
}
\seealso{
\code{\link[Ramses]{create_therapy_episodes}()}
}
