% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{longitudinal_table}
\alias{longitudinal_table}
\alias{longitudinal_table,TherapyEpisode-method}
\alias{longitudinal_table,MedicationRequest-method}
\alias{longitudinal_table,Encounter-method}
\title{Get the longitudinal_table}
\usage{
longitudinal_table(object, collect = FALSE)

\S4method{longitudinal_table}{TherapyEpisode}(object, collect = FALSE)

\S4method{longitudinal_table}{MedicationRequest}(object, collect = FALSE)

\S4method{longitudinal_table}{Encounter}(object, collect = FALSE)
}
\arguments{
\item{object}{an object of class \code{Encounter}, \code{TherapyEpisode}, 
or \code{MedicationRequest}}

\item{collect}{if \code{TRUE}, collect the remote \code{tbl_sql} and return a local 
\code{tbl_df}. The default is \code{FALSE}, and simply returns the remote \code{tbl_sql}}
}
\value{
an object of class \code{tbl}
}
\description{
Get the longitudinal_table
}
