% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{map_charlson_comorbidities}
\alias{map_charlson_comorbidities}
\title{Map ICD-10 to Charlson comorbidities and weights}
\source{
Adapted from \code{\link[comorbidity]{comorbidity}()}
}
\usage{
map_charlson_comorbidities(df, icd_column)
}
\arguments{
\item{df}{a data frame containing ICD-10 look up data}

\item{icd_column}{a variable name or column number in `df` 
containing ICD-10 codes.}
}
\value{
The input data frame `df` enhanced with the following 
variables:
  \itemize{
    \item \code{comorb} a character vector of 17 comorbidity 
    codes (adapted from the 
    \code{\link[comorbidity]{comorbidity}} package)
    \item \code{comorb_group} a character vector of codes 
    grouping \code{comorb} codes into 14 categories 
    collapsing different severities of the same condition 
    (for diabetes, liver disease, cancer). This is to avoid 
    double counting comorbidities.
    \item \code{charlson_weights} an integer vector of weights 
    from \insertCite{Quan2005}{Ramses} used to compute the 
    weighted Charlson comorbidity score
  }
}
\description{
Map ICD-10 or ICD-10-CM codes to a list of chronic
comorbidities, comorbidity groups, and Charlson Comorbidity Index 
weights from \insertCite{Quan2005}{Ramses}.
}
\examples{
mock_icd_data <- data.frame(
   list(icd10_code = c("J44", "J44X", "J440")), 
   stringsAsFactors = FALSE)
mock_icd_data <- map_charlson_comorbidities(
  mock_icd_data, "icd10_code")
mock_icd_data
}
\references{
\insertAllCited{}
}
