% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{map_ICD10_CCS}
\alias{map_ICD10_CCS}
\alias{map_ICD10_CCSR}
\title{Map ICD-10 codes to the Clinical Classifications Software (CCS)}
\usage{
map_ICD10_CCS(df, icd_column)

map_ICD10_CCSR(df, icd_column)
}
\arguments{
\item{df}{a data frame containing ICD-10 codes}

\item{icd_column}{a variable name or column number in \code{df} 
containing ICD-10 codes.}
}
\value{
The input data frame with additional CCS/CCSR classifications.
CCS categories are mutually exclusive. CCSR categories are not.
}
\description{
Map ICD-10 codes to clinically meaningful 
categories within the Clinical Classifications Software (\code{\link{ccs}})
and the Clinical Classifications Software Refined (\code{\link{ccsr}}).
}
\details{
This function is designed to be compatible with 
both ICD-10 and ICD-10-CM versions.
}
\examples{
mock_icd_data <- data.frame(
  list(icd10_code = c("J44", "J44X", "J440")), 
  stringsAsFactors = FALSE)
mock_icd_data <- map_ICD10_CCS(mock_icd_data, "icd10_code")
mock_icd_data <- map_ICD10_CCSR(mock_icd_data, "icd10_code")
str(mock_icd_data)
}
\seealso{
\code{\link{ccs}}, \code{\link{ccsr}}.
}
