% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{map_infections_abx_indications}
\alias{map_infections_abx_indications}
\title{Map ICD-10 to common infections and antibiotic indications}
\usage{
map_infections_abx_indications(df, icd_column)
}
\arguments{
\item{df}{a data frame containing ICD-10 codes}

\item{icd_column}{a variable name or column number in \code{df} 
containing ICD-10 codes.}
}
\value{
The input data frame \code{x} with additional variables 
describing infection type and whether antibiotics are indicated.
}
\description{
Map ICD-10 or ICD-10-CM codes by infection types 
and whether antibiotics are commonly indicated based on definitions
set by \insertCite{Hashimoto2020}{Ramses}.
}
\details{
For more detail on infection categories and indications
for antibiotic therapy, see \code{\link{antibiotic_icd_indications}}.

This function is designed to be compatible with both ICD-10 and 
ICD-10-CM versions.
}
\examples{
mock_icd_data <- data.frame(
   list(icd10_code = c("J44", "J44X", "J440")), 
   stringsAsFactors = FALSE)
mock_icd_data <- map_infections_abx_indications(
  mock_icd_data, "icd10_code")
mock_icd_data
}
\seealso{
\code{\link{antibiotic_icd_indications}}
}
