% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{MedicationRequest-class}
\alias{MedicationRequest-class}
\alias{MedicationRequest}
\title{An S4 class to represent a drug-dose medication request}
\usage{
MedicationRequest(conn, id)
}
\arguments{
\item{conn}{a database connection}

\item{id}{a prescription identifier}
}
\description{
This class represent a clinician's request for a single 
drug-dose order. as monotherapy a combination therapy
}
\section{Slots}{

\describe{
\item{\code{id}}{a prescription identifier}

\item{\code{conn}}{a database connection}

\item{\code{record}}{a \code{tbl_sql} for the corresponding database record}
}}

