% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{parenteral_changes_get}
\alias{parenteral_changes_get}
\title{Title}
\usage{
parenteral_changes_get(therapy_episode)
}
\arguments{
\item{therapy_episode}{}
}
\value{
a list containing positions of matches. These are vectors of 
length 2, containing the sequence starting and finishing positions
}
\description{
Title
}
\details{
This function returns positions of sequences of zeroes followed by a 1
or a terminating sequence of zeros only.
Periods of up to 9 hours without therapy (e.g. between two prescriptions)
are ignored.
This function returns positions of sequences of zeroes followed by a 1
or a terminating sequence of zeros only.
Periods of up to 9 hours without therapy (e.g. between two prescriptions)
are ignored.
The logics are intended to disregard short sequences of ones (like a 
very short administration of oral only, for up to 9 hours, in between
two prescriptions of IV)
}
