% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_records.R
\docType{data}
\name{patients}
\alias{patients}
\title{Simulated patients}
\format{
A data frame of patient demographics.
}
\source{
Simulated data
}
\usage{
patients
}
\description{
This simulated dataset contains records of 200 patients 
corresponding to hospitalisation records in \link{inpatient_episodes}.
}
\keyword{datasets}
