% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ramses-package.R
\docType{package}
\name{Ramses-package}
\alias{Ramses}
\alias{Ramses-package}
\title{Ramses: R Package for Antimicrobial Stewardship & Surveillance}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

A one-stop shop for antimicrobial stewardship data warehousing for hospitals. 
    Ramses supports the modelling and analysis of hospital electronic health records 
    (episode of care, prescriptions, administration, pathology and investigations) to help
    keep antibiotics working for years to come.
}
\details{
R package for Antimicrobial Stewardship & Surveillance
   
   Copyright (C) 2020 University College London
   
   This program is free software; you can redistribute it and/or modify
   it under the terms of the GNU General Public License as published by
   the Free Software Foundation; either version 3 of the License, or
   (at your option) any later version.
   
   This program is distributed in the hope that it will be useful,
   but WITHOUT ANY WARRANTY; without even the implied warranty of
   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
   GNU General Public License for more details.
}
\seealso{
Preserving Antibiotics through Safe Stewardship (PASS) research 
project: \url{https://pass-antibiotics.github.io}
}
\author{
\strong{Maintainer}: Peter Dutey-Magni \email{p.dutey-magni@ucl.ac.uk} (\href{https://orcid.org/0000-0002-8942-9836}{ORCID}) [researcher]

Other contributors:
\itemize{
  \item Laura Shallcross \email{l.shallcross@ucl.ac.uk} (\href{https://orcid.org/0000-0003-1713-2555}{ORCID}) [research team head]
  \item University College London [copyright holder]
}

}
\keyword{internal}
