% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{RamsesObject-class}
\alias{RamsesObject-class}
\title{An S4 virtual class for Ramses objects}
\description{
An S4 virtual class for Ramses objects
}
\section{Slots}{

\describe{
\item{\code{id}}{a character identifier}

\item{\code{conn}}{a database connection}

\item{\code{record}}{a \code{tbl_sql} for the corresponding database record}
}}

