% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\docType{data}
\name{reference_aware}
\alias{reference_aware}
\title{English Adaptation of the WHO AWaRe Index}
\format{
A data frame with 208 antimicrobial agents, referenced 
against the World Health Organisation's Anatomical Therapeutic Chemical (ATC) 
classification \insertCite{WHO-ATC2020}{Ramses}, 
Public Health England's AWaRe Indices, 
and the SNOMED CT medical product concept codes and names.
It contains 6 columns:
\describe{
\item{\code{ATC_code}}{ATC code of the antibiotic}
\item{\code{ATC_route}}{route of administration as defined in
the ATC ("O" = oral; "P" = parenteral; "R" = rectal;
"V" = vaginal)}
\item{\code{ATC_name}}{ATC name of the antibiotic}
\item{\code{aware_category}}{AWaRe index ("Access", "Watch", "Reserve")}
\item{\code{version}}{AWaRe Index version (eg. "WHO", "England")}
\item{\code{year}}{AWaRe Index version year}
\item{\code{VTM_code}}{SNOMED CT medicinal product code/Virtual Therapeutic Moiety 
code in the SNOMED CT UK Drug Extension \insertCite{SNOMEDDrugUK2020}{Ramses}}
\item{\code{VTM_name}}{SNOMED CT medicinal product code/Virtual Therapeutic Moiety 
name in the SNOMED CT UK Drug Extension \insertCite{SNOMEDDrugUK2020}{Ramses}}
}
}
\source{
Adapted from \insertCite{WHO-aware2019,PHE-aware2019;textual}{Ramses} with Virtual  
Therapeutic Moiety equivalents from the NHS Dictionary of Medicines (dm+d)
\insertCite{SNOMEDDrugUK2020}{Ramses}.
}
\usage{
reference_aware
}
\description{
This `AWaRe Index' is a classification by the World Health
Organisation of key antibiotics into three categories
\insertCite{Sharland2018}{Ramses} to:
\itemize{
       \item improve access (Access)
       \item monitor important antibiotics (Watch) and 
       \item preserve effectiveness of ‘last resort’ antibiotics (Reserve).
}
It was adapted by Public Health England for antibiotic stewardship
and quality improvement in English hospitals 
\insertCite{PHE-aware2019}{Ramses}.

The version included in Ramses contains an additional row for metronidazole
J01XD01 administered orally (classified as Access). This is sometimes 
used in combination with other antibacterials that do not cover anaerobic 
organisms, but is not present in the Anatomical Therapeutic Chemical (ATC) 
classification.
}
\examples{
data("reference_aware")
head(reference_aware)
}
\references{
{\insertAllCited{}}
}
\keyword{datasets}
