% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\docType{data}
\name{reference_drug_frequency}
\alias{reference_drug_frequency}
\title{Prescription frequency abbreviation look-up table}
\format{
A data frame with 64 abbreviations and their corresponding daily frequency value.
}
\usage{
reference_drug_frequency
}
\description{
This look-up table provides the number of times a drug should 
be administered per day based on the prescription frequency abbreviation 
(including latin abbreviations).

Example: BD stands for \emph{bis die}, meaning twice a day. The 
corresponding `daily_frequency`` value is 2.
}
\keyword{datasets}
