% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\docType{data}
\name{reference_specimen_type}
\alias{reference_specimen_type}
\title{Laboratory specimen type reference table (SNOMED CT concepts)}
\format{
A data frame.
}
\usage{
reference_specimen_type
}
\description{
This reference table contains the list of admissible specimen
types in microbiology records (see \code{\link[Ramses]{validate_microbiology}()}).
These concepts are from the 
\href{SNOMED CT International Edition}{https://browser.ihtsdotools.org/} and 
correspond to the descendants of the \code{123038009 | Specimen (specimen) |} 
concept.
}
\details{
This data frame contains 5 variables:
\describe{
    \item{\code{conceptId}}{SNOMED CT concept code}
    \item{\code{moduleId}}{SNOMED CT module code}
    \item{\code{fsn_term}}{SNOMED CT Full Specified Name}
    \item{\code{pt_term}}{SNOMED CT Preferred Term}
    \item{\code{snomed_release_version}}{SNOMED CT RF2 release date}
}
}
\examples{
data("reference_specimen_type")
head(reference_specimen_type)
}
\seealso{
\url{https://snomedizer.web.app}
}
\keyword{datasets}
