% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{show,RamsesObject-method}
\alias{show,RamsesObject-method}
\alias{show,Encounter-method}
\alias{show,TherapyEpisode-method}
\alias{show,MedicationRequest-method}
\title{Show a Ramses object}
\usage{
\S4method{show}{RamsesObject}(object)

\S4method{show}{Encounter}(object)

\S4method{show}{TherapyEpisode}(object)

\S4method{show}{MedicationRequest}(object)
}
\arguments{
\item{RamsesObject}{an object of class RamsesObject}
}
\value{
show returns an invisible \code{NULL}.
}
\description{
Print a summary of a Ramses object.
}
\seealso{
\code{\link{methods}{show}}
}
