% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ramses-deprecated.R
\name{therapy_table}
\alias{therapy_table}
\title{Get the therapy table (DEPRECATED)}
\usage{
therapy_table(object, collect = FALSE)
}
\arguments{
\item{object}{an object of class \code{TherapyEpisode}}

\item{collect}{if \code{TRUE}, collect the remote \code{tbl_sql} and return a local 
\code{tbl_df}. The default is \code{FALSE}, and simply returns the remote \code{tbl_sql}}
}
\value{
an object of class \code{tbl}
}
\description{
This function is now deprecated. Please use
 \code{\link{longitudinal_table}()}.
}
\seealso{
\code{\link{Ramses-deprecated}}
}
