% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{therapy_table}
\alias{therapy_table}
\alias{therapy_table,TherapyEpisode-method}
\alias{therapy_table,MedicationRequest-method}
\title{Get the therapy table}
\usage{
therapy_table(object, collect = FALSE)

\S4method{therapy_table}{TherapyEpisode}(object, collect = FALSE)

\S4method{therapy_table}{MedicationRequest}(object, collect = FALSE)
}
\arguments{
\item{object}{an object of class \code{TherapyEpisode}}

\item{collect}{if \code{TRUE}, collect the remote \code{tbl_sql} and return a local 
\code{tbl_df}. The default is \code{FALSE}, and simply returns the remote \code{tbl_sql}}
}
\value{
an object of class \code{tbl}
}
\description{
Get the therapy table
}
