% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{therapy_table_get}
\alias{therapy_table_get}
\alias{therapy_table_get,TherapyEpisode-method}
\title{Get the therapy table}
\usage{
therapy_table_get(object, collect = FALSE)

\S4method{therapy_table_get}{TherapyEpisode}(object, collect = FALSE)
}
\arguments{
\item{object}{an object of class \code{TherapyEpisode}}

\item{collect}{whether to collect the remote \code{tbl_sql} and return a local 
\code{tbl_df}. The default is \code{FALSE}}
}
\value{
an object of class \code{tbl}
}
\description{
Get the therapy table
}
