% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline.R
\name{therapy_timeline}
\alias{therapy_timeline}
\title{Build an antiinfective timeline visualisation}
\usage{
therapy_timeline(
  conn,
  patient_identifier,
  date1 = NA,
  date2 = NA,
  load_timevis_dependencies = FALSE
)
}
\arguments{
\item{conn}{a database connection}

\item{patient_identifier}{a length-one character vector for a patient ID}

\item{date1}{(optional) a date window minimum to focus the timeline on by default}

\item{date2}{(optional) a date window maximum to focus the timeline on by default}

\item{load_timevis_dependencies}{a boolean indicating whether jQuery 
and Boostrap should be loaded (default is FALSE). See \link[timevis]{timevis}
for detail}
}
\value{
a timevis object
}
\description{
Build an antiinfective timeline visualisation
}
