% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{TherapyEpisode-class}
\alias{TherapyEpisode-class}
\alias{TherapyEpisode}
\alias{TherapyEpisode.DBIConnection}
\alias{TherapyEpisode.RamsesObject}
\title{An S4 class to represent therapy episodes}
\usage{
TherapyEpisode(...)

TherapyEpisode.DBIConnection(conn, id)

TherapyEpisode.RamsesObject(object)
}
\arguments{
\item{...}{generic signature}

\item{conn}{a database connection}

\item{id}{a vector of one or several therapy episode identifiers}

\item{object}{an object of class \code{MedicationRequest} or 
\code{Prescription}}
}
\description{
An S4 class to represent therapy episodes
}
\section{Slots}{

\describe{
\item{\code{id}}{a character therapy episode identifier}

\item{\code{conn}}{a database connection}

\item{\code{record}}{a \code{tbl_sql} for the corresponding database record}

\item{\code{therapy_table}}{a \code{tbl_sql} for the longitudinal therapy table}
}}

