% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{transitive_closure_control}
\alias{transitive_closure_control}
\title{Set parameters for building combination therapy identifiers and
therapy episode identifiers}
\usage{
transitive_closure_control(
  max_continuation_gap = 36,
  max_combination_authoring_gap = 6,
  max_combination_start_gap = 24
)
}
\arguments{
\item{max_continuation_gap}{a positive integer setting the maximum number of hours
tolerated between the end of a prescription and the start of a subsequent 
prescription in the same therapy episode. The default is 36 hours.}

\item{max_combination_authoring_gap}{a positive integer setting the maximum number
of hours tolerated between the time of authoring of two prescriptions 
administered as combination therapy. The default is 6 hours.}

\item{max_combination_start_gap}{a positive integer setting the maximum number of 
hours tolerated between the start of administration of two prescriptions
administred as combination therapy. The default is 24 hours.}
}
\value{
A list of three positive integer variables.
}
\description{
Set parameters controlling how prescriptions are 
linked into combinations and therapy episodes
}
\details{
Electronic prescribing and administration systems do not usually 
define:
\itemize{
  \item combination therapy: prescribing two or more antimicrobials  
  concurrently for the same indication
  \item therapy episodes: the set of antimicrobial prescriptions  
  administered consecutively or concurrently and corresponding to 
  a single indication and intent. 
}

Because of this, \code{\link{load_medications}}() first examines all 
prescriptions authored in a given hospital admission to ascertain such 
links between prescriptions. This involves looking at patterns of 
overlap and succession of prescriptions, and using a graph theory method 
known as 'transitive closure'.
}
\examples{
# default parameters
transitive_closure_control(
  max_continuation_gap = 36,
  max_combination_authoring_gap = 6,
  max_combination_start_gap = 24)
}
