% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_data.R
\name{validate_administrations}
\alias{validate_administrations}
\title{Validate medication administration records}
\usage{
validate_administrations(data)
}
\arguments{
\item{data}{a data frame containing one row per prescription. See Details for other conditions.}
}
\value{
NULL if the dataset is valid
}
\description{
Validate medication administration records
}
\details{
In order to pass the validation, prescription data should contain the following columns with 
  no missing data:
  \itemize{
     \item \code{patient_id}: a patient identifier with no missing value
     \item \code{prescription_id}: a prescription identifier with no missing value
     \item \code{administration_id}: an administration identifier with no missing value
     \item \code{drug_id}: identifier of the drug (from a dictionary such as SNOMED-CT or
      from \code{\link{AMR}{as.ab}()})
     \item \code{drug_name}: preferred name of the drug in the drug dictionary
     \item \code{drug_display_name}: drug name to display in reports (can be the same as \code{drug_name})
     \item \code{ATC_code}
     \item \code{ATC_group}
     \item \code{ATC_route}
     \item \code{administration_time}: timestamp of the drug administration
     \item \code{administration_context}: either 'inpatient', 'opat', 'discharge'
    }
   
  Optional columns include:
  \itemize{
     \item \code{dose}: dose administered
     \item \code{unit}: dose unit, see \code{\link[units]{ud_units}}
     \item \code{route}: route of administration of the drug
     \item \code{ddd}: the number of defined daily doses, see \code{\link{compute_DDDs}()}.
  }
}
