% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_data.R
\name{validate_inpatient_diagnoses}
\alias{validate_inpatient_diagnoses}
\title{Validate inpatient diagnosis records}
\usage{
validate_inpatient_diagnoses(diagnoses_data, diagnoses_lookup)
}
\arguments{
\item{diagnoses_data}{a data frame containing clinical diagnoses, with, at minimum, 
variables `patient_id`, `spell_id`, `episode_number`, `icd_code`, `diagnosis_position`}

\item{diagnoses_lookup}{a data frame containing an ICD-10 reference look up table}
}
\value{
a boolean indicating whether the data passed the set of validations
}
\description{
Validate contraints on diagnosis records, namely that the minimum 
variables are present, and that all `icd_code` values can be looked up in an 
ICD-10 reference table
}
\examples{
lookup_icd <- dplyr::distinct(Ramses::inpatient_diagnoses, icd_code)
lookup_icd <- dplyr::filter(lookup_icd, !is.na(icd_code))
lookup_icd$icd_label <- "ICD-10 code label text"
validate_inpatient_diagnoses(Ramses::inpatient_diagnoses, lookup_icd)
}
