% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_data.R
\name{validate_inpatient_episode_dates}
\alias{validate_inpatient_episode_dates}
\title{Validate the consistency of inpatient episode dates}
\usage{
validate_inpatient_episode_dates(data)
}
\arguments{
\item{data}{a data frame object}
}
\value{
A logical value indicating success
}
\description{
Validate the consistency of inpatient episode dates
}
