% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_data.R
\name{validate_inpatient_episodes}
\alias{validate_inpatient_episodes}
\title{Validate inpatient episode records}
\usage{
validate_inpatient_episodes(data)
}
\arguments{
\item{data}{a data frame containing}
}
\value{
TODO
}
\description{
Validate inpatient episode records
}
