% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_data.R
\name{validate_inpatient_spells}
\alias{validate_inpatient_spells}
\title{Validate the consistency of admission and discharge dates}
\usage{
validate_inpatient_spells(data)
}
\arguments{
\item{data}{a data frame object}
}
\description{
Validate the consistency of admission and discharge dates
}
