% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_data.R
\name{validate_microbiology}
\alias{validate_microbiology}
\title{Validate microbial isolates & susceptibility records}
\usage{
validate_microbiology(specimens, isolates, susceptibilities)
}
\arguments{
\item{specimens}{a data frame with one row per specimen sent
to laboratory (see details)}

\item{isolates}{a data frame with one row per microorganism
isolated from the laboratory specimen (see details)}

\item{susceptibilities}{a data frame with one row per susceptibility
(see details)}
}
\value{
TRUE if the validation is passed
}
\description{
Validate microbial isolates & susceptibility records
}
\section{\code{specimens} data frame}{

The following fields are mandatory:
\describe{
   \item{\code{specimen_id}}{a unique identifier with no missing value}
   \item{\code{patient_id}}{a patient identifier with no missing value}
   \item{\code{spell_id}}{a hospital spell identifier (if the specimen was 
   sampled during admission)}
   \item{\code{status}}{one value from the following 
   \href{https://www.hl7.org/fhir/valueset-specimen-status.html}{FHIR R4}
     reference set: \itemize{
       \item \code{"available"}: The physical specimen is present and 
       in good condition.
       \item \code{"unavailable"}: There is no physical specimen because it is either 
       lost, destroyed or consumed.
       \item \code{"unsatisfactory"}: The specimen cannot be used because of a
        quality issue such as a broken container, contamination, or too old.	
       \item \code{"entered-in-error"}: The specimen was entered in error and 
       therefore nullified.
     }}
   \item{\code{specimen_datetime}}{datetime when specimen was sampled or 
   received for processing.}
   \item{\code{specimen_type_code}}{}
   \item{\code{specimen_type_name}}{}
   \item{\code{specimen_type_display}}{}
}

The following fields are optional:
\describe{
   \item{\code{a}}{a}
   \item{\code{a}}{a}
   \item{\code{a}}{a}
 }


The following fields are mandatory:
\describe{
   \item{\code{a}}{a}
   \item{\code{a}}{a}
   \item{\code{a}}{a}
 }

The following fields are optional:
\describe{
   \item{\code{a}}{a}
   \item{\code{a}}{a}
   \item{\code{a}}{a}
}
}

\section{\code{susceptibilities} data frame}{

The following fields are mandatory:
\describe{
  \item{specimen_id}{a specimen identifier with no missing value}
  \item{status}{one value from the following 
     \href{https://www.hl7.org/fhir/valueset-specimen-status.html}{FHIR reference set}:
     \itemize{
        \item \code{"available"}: The physical specimen is present and in good condition.
        \item \code{"unavailable"}: There is no physical specimen because it is either lost, destroyed or consumed.
        \item \code{"unsatisfactory"}: The specimen cannot be used because of a quality issue such as a broken container, contamination, or too old.
        \item \code{"entered-in-error"}: The specimen was entered in error and therefore nullified.
     }}
}

The following fields are optional:
\describe{
  \item{blah}{blah}
}
}

