% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_data.R
\name{validate_prescriptions}
\alias{validate_prescriptions}
\title{Validate medication prescription records}
\usage{
validate_prescriptions(data)
}
\arguments{
\item{data}{a data frame containing one row per prescription. See Details for other conditions.}
}
\value{
NULL if the dataset is valid
}
\description{
Validate medication prescription records
}
\details{
In order to pass the validation, prescription data should contain the following columns with 
  no missing data:
  \itemize{
     \item \code{patient_id}: a patient identifier with no missing value
     \item \code{prescription_id}: a prescription identifier with no missing value
     \item \code{drug_id}: identifier of the drug (from a dictionary such as SNOMED-CT or
      from \code{\link{AMR}{as.ab}()})
     \item \code{drug_name}: preferred name of the drug in the drug dictionary
     \item \code{drug_display_name}: drug name to display in reports (can be the same as \code{drug_name})
     \item \code{ATC_code}
     \item \code{ATC_group}
     \item \code{ATC_route}
     \item \code{authoring_date}: timestamp when the prescription was issued
     \item \code{prescription_start}: timestamp of the prescription start
     \item \code{prescription_end}: timestamp of the prescription end
     \item \code{prescription_context}: either 'inpatient', 'opat', 'discharge'
    }
   
  Optional columns include:
  \itemize{
     \item \code{ddd}: the number of defined daily doses, see \code{\link{compute_DDDs}()}.
  }
}
