% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{bridge_tables}
\alias{bridge_tables}
\alias{bridge_episode_prescription_overlap}
\alias{bridge_episode_prescription_initiation}
\alias{bridge_spell_therapy_overlap}
\title{Create database bridge tables}
\usage{
bridge_tables(conn, overwrite = FALSE, silent = FALSE)

bridge_episode_prescription_overlap(conn, overwrite = FALSE)

bridge_episode_prescription_initiation(conn, overwrite = FALSE)

bridge_spell_therapy_overlap(conn, overwrite = FALSE)
}
\arguments{
\item{conn}{a database connection}

\item{overwrite}{if \code{TRUE}, will overwrite any existing
database table. The default is \code{FALSE}}

\item{silent}{a boolean indicating whether the function should be
executed without progress bar. Default is \code{FALSE}.}
}
\value{
TRUE tables were successfully created
}
\description{
Create or re-create bridge tables to facilitate linking prescribing
events with spells or episodes of care. Bridge tables are used when computing
rates of prescribing per admission or per 1,000 bed-days.
}
\details{
\describe{
   \item{\code{bridge_tables()}}{Generate all tables below.}
   \item{\code{bridge_episode_prescription_overlap()}}{Links prescriptions
   to inpatient episodes during which they were administered. The resulting 
   table is the natural join of \code{inpatient_episodes} and 
   \code{drug_prescriptions} based on a match on the patient identifier 
   and a time overlap between prescriptions and inpatient episodes.}
   \item{\code{bridge_episode_prescription_initiation()}}{Links prescriptions
   with the inpatient episode when they were authored. The resulting table 
   differs from {\code{bridge_spell_therapy_overlap}}: it links prescriptions 
   to the episode referencing the clinical team who prescribed them, rather
   that episodes during which the prescription was administered. The resulting
   table is the natural join of \code{inpatient_episodes} and 
   \code{drug_prescriptions} based on a match on the patient identifier
   and the prescription authoring date being comprised before the episode 
   start and end dates.}
   \item{\code{bridge_spell_therapy_overlap()}}{Links therapy episodes
   to inpatient spells during which they were administered. The resulting 
   table is the natural join of \code{inpatient_episodes} and 
   \code{drug_prescriptions} based on a match on the patient identifier 
   and a time overlap between prescriptions and dates of admission and discharge.}
} 
\code{inpatient_episodes}
}
