% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{bridge_tables}
\alias{bridge_tables}
\alias{bridge_episode_prescription_overlap}
\alias{bridge_episode_prescription_initiation}
\alias{bridge_spell_therapy_overlap}
\title{Create database bridge tables}
\usage{
bridge_tables(conn, overwrite = FALSE, silent = FALSE)

bridge_episode_prescription_overlap(conn, overwrite = FALSE)

bridge_episode_prescription_initiation(conn, overwrite = FALSE)

bridge_spell_therapy_overlap(conn, overwrite = FALSE)
}
\arguments{
\item{conn}{a database connection}

\item{overwrite}{if \code{TRUE}, will overwrite any existing
database table. The default is \code{FALSE}}

\item{silent}{if \code{TRUE}, the progress bar will be hidden. The default is 
\code{FALSE}.}
}
\value{
\code{TRUE} if tables were successfully created
}
\description{
Create or re-create bridge tables to facilitate linking prescribing
events with spells or episodes of care. Bridge tables are used when computing
rates of prescribing per admission or per 1,000 bed-days.
}
\details{
Prescriptions with status \code{"entered-in-error"}, \code{"draft"}, 
\code{"cancelled"}, or \code{"unknown"} are not taken into account.
\describe{
   \item{\code{bridge_tables()}}{Generates all bridge tables.}
   \item{\code{bridge_episode_prescription_overlap()}}{Links prescriptions
   with inpatient episodes when they were administered. The resulting 
   table is the natural join of \code{inpatient_episodes} and 
   \code{drug_prescriptions} based on matching patient identifiers 
   and a time overlap between prescriptions and inpatient episodes.}
   \item{\code{bridge_episode_prescription_initiation()}}{Links prescriptions
   with inpatient episodes when they were authored. The resulting table 
   differs from {\code{bridge_episode_prescription_overlap}}: it links prescriptions 
   to the episode referencing the clinical team who prescribed them, rather
   that episodes during which the prescription was administered. The resulting
   table is the natural join of \code{inpatient_episodes} and 
   \code{drug_prescriptions} based on matching patient identifiers 
   and the prescription authoring date being comprised between the episode 
   start and end dates.}
   \item{\code{bridge_spell_therapy_overlap()}}{Links therapy episodes
   with inpatient spells during which they were administered. The resulting 
   table is the natural join of \code{inpatient_episodes} and 
   \code{drug_therapy_episodes} based on matching patient identifiers 
   and a time overlap between therapy episodes and hospital stays.}
} 
\code{inpatient_episodes}
}
