% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_features.R
\name{clinical_feature_interval}
\alias{clinical_feature_interval}
\alias{clinical_feature_interval,TherapyEpisode-method}
\title{Therapy table feature: number of clinical observations falling in an interval}
\usage{
clinical_feature_interval(
  x,
  observation_intervals,
  hours,
  observation_code_system = NULL
)

\S4method{clinical_feature_interval}{TherapyEpisode}(
  x,
  observation_intervals,
  hours,
  observation_code_system = NULL
)
}
\arguments{
\item{x}{a \code{\link{TherapyEpisode}} object}

\item{observation_intervals}{a named list of numeric vectors of length 1 
(for a threshold) or 2 (for an interval). Names of vectors must match the
 \code{observation_code} field in the \code{inpatient_investigation}
table (see \code{\link{validate_investigations}()}. See example}

\item{hours}{the maximum number of hours the observation should date back from
\code{t_start}, the starting time of every row in the therapy table}

\item{observation_code_system}{(optional, reserved to situations where 
\code{observation_code} is ambiguous across code systems/vocabularies) a single 
character string specifying the code system identifier (example:
 \code{"http://snomed.info/sct"}) of \code{observation_code}.

The default (\code{NULL}) filters observations using the \code{observation_code} only.}
}
\value{
an object of class \code{\link{TherapyEpisode}}
}
\description{
Add clinical features (variables) to a therapy episode table 
containing the number of observations falling (a) above/below a given threshold
or (b) inside/outside a given interval.
}
\details{
The feature will be computed exclusively on numeric investigations 
marked with status \code{"final"}, \code{"preliminary"}, \code{"corrected"}, 
or \code{"amended"}.
}
\examples{
\dontrun{
temperature_check <- clinical_feature_interval(
   TherapyEpisode(conSQLite, "4d611fc8886c23ab047ad5f74e5080d7"),
   observation_intervals = list("8310-5" = c(36, 38)),
   hours = 24
   )
str(therapy_table(temperature_check, collect = TRUE))
}
}
