% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_features.R
\name{clinical_feature_mean}
\alias{clinical_feature_mean}
\alias{clinical_feature_mean,TherapyEpisode-method}
\title{Therapy table feature: running mean value of a clinical observation}
\usage{
clinical_feature_mean(
  x,
  observation_code,
  hours,
  observation_code_system = NULL,
  compute = TRUE
)

\S4method{clinical_feature_mean}{TherapyEpisode}(
  x,
  observation_code,
  hours,
  observation_code_system = NULL,
  compute = TRUE
)
}
\arguments{
\item{x}{a \code{\link{TherapyEpisode}} object}

\item{observation_code}{a character vector of clinical investigation codes
matching the \code{observation_code} field in the \code{inpatient_investigation}
table (see \code{\link{validate_investigations}()}}

\item{hours}{the maximum number of hours the observations included in the mean
should date back from \code{t_start}, the starting time of every row 
in the therapy table}

\item{observation_code_system}{(optional, reserved to situations where 
\code{observation_code} is ambiguous across code systems/vocabularies) a single 
character string specifying the code system identifier (example:
 \code{"http://snomed.info/sct"}) of \code{observation_code}.

The default (\code{NULL}) filters observations using the \code{observation_code} only.}

\item{compute}{if \code{TRUE} (the default), the remote therapy table will 
be computed on the remote server. This is generally faster.}
}
\value{
an object of class \code{\link{TherapyEpisode}}
}
\description{
Add a clinical feature (variable) to a therapy episode table 
containing the arithmetic mean of clinical observations of interest
}
\details{
The feature will be computed exclusively on numeric investigations 
marked with status \code{"final"}, \code{"preliminary"}, \code{"corrected"}, 
or \code{"amended"}.
}
\examples{
\dontrun{
conSQLite <- create_mock_database("example.sqlite")
temperature_check <- clinical_feature_mean(
   TherapyEpisode(conSQLite, "4d611fc8886c23ab047ad5f74e5080d7"),
   observation_code = "8310-5",
   hours = 24
   )
str(therapy_table(temperature_check, collect = TRUE))
}
}
