% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_features.R
\name{clinical_feature_ols_trend}
\alias{clinical_feature_ols_trend}
\alias{clinical_feature_ols_trend,TherapyEpisode-method}
\title{Therapy table feature: temporal trend of clinical observations}
\usage{
clinical_feature_ols_trend(
  x,
  observation_code,
  hours,
  observation_code_system = NULL
)

\S4method{clinical_feature_ols_trend}{TherapyEpisode}(
  x,
  observation_code,
  hours,
  observation_code_system = NULL
)
}
\arguments{
\item{x}{a \code{\link{TherapyEpisode}} object}

\item{observation_code}{a character vector of clinical investigation codes
matching the \code{observation_code} field in the \code{inpatient_investigation}
table (see \code{\link{validate_investigations}()}}

\item{hours}{the maximum number of hours the observations should date back from
\code{t_start}, the starting time of every row in the therapy table}

\item{observation_code_system}{(optional) reserved to  instance where 
\code{observation_code} is ambiguous and does not uniquely identify 
observations across code systems (vocabularies), specify a single
code system identifier (eg \code{"http://snomed.info/sct"}) to use.
The default is \code{NULL} and will only filter observation using
\code{observation_code}.}
}
\value{
an object of class \code{\link{TherapyEpisode}}
}
\description{
Add a clinical feature to a therapy episode table corresponding
to the Ordinary Least Squares (OLS) intercept and slope of clinical
observations made until table time \code{t}
}
\details{
NOTE: only numeric investigations marked with status \code{"final"}, 
\code{"preliminary"}, \code{"corrected"}, or \code{"amended"} will be sourced.

The returned regression slope coefficient corresponds to the mean change 
associated with a 1-hour time increment.

The returned regression intercept is defined with respect to time equals
zero at \code{t_start}. It thus corresponds to the linear (straight line) 
extrapolation of the trend to \code{t_start}.
}
\examples{
\dontrun{
temperature_check <- clinical_feature_ols_trend(
   TherapyEpisode(conSQLite, "4d611fc8886c23ab047ad5f74e5080d7"),
   observation_code = "8310-5",
   hours = 24
   )
str(therapy_table(temperature_check, collect = TRUE))
}
}
