% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{collect}
\alias{collect}
\alias{collect,TherapyEpisode-method}
\alias{collect,Encounter-method}
\alias{collect,RamsesObject-method}
\title{Retrieve the database record of a Ramses Object}
\usage{
collect(x, ...)

\S4method{collect}{TherapyEpisode}(x)

\S4method{collect}{Encounter}(x)

\S4method{collect}{RamsesObject}(x)
}
\arguments{
\item{x}{an object of class \code{RamsesObject}}

\item{...}{arguments passed on to methods (compatibility with \code{dplyr}'s
S3 generic)}
}
\value{
a data frame of class \code{tbl_df}
}
\description{
Retrieve a data frame of the record belonging to a Ramses 
object instance.
}
\seealso{
\code{\link[Ramses]{longitudinal_table}()}
}
