% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{collect_ramses_tbl}
\alias{collect_ramses_tbl}
\title{Collect SQL tibble with correct date-time specifications}
\usage{
collect_ramses_tbl(tbl)
}
\arguments{
\item{tbl}{a `tbl_sql` object}
}
\value{
a `tbl_df` object
}
\description{
This wrapper function for \link[dplyr]{collect} will convert
relevant character columns to \code{Date} and \code{POSIXct} type when collecting 
SQLite tables. This addresses the absence of date and time data types
in SQLite. Tables from other relational database systems are note affected.
}
\seealso{
\url{https://www.sqlite.org/datatype3.html#date_and_time_datatype}
}
