% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{compute}
\alias{compute}
\alias{compute,TherapyEpisode-method}
\alias{compute,Encounter-method}
\alias{compute,RamsesObject-method}
\title{Compute the database records of a Ramses object}
\usage{
compute(x, ...)

\S4method{compute}{TherapyEpisode}(x)

\S4method{compute}{Encounter}(x)

\S4method{compute}{RamsesObject}(x)
}
\arguments{
\item{x}{an object of class \code{RamsesObject}}

\item{...}{arguments passed on to methods (compatibility with \code{dplyr}'s
S3 generic)}
}
\value{
an object of class \code{RamsesObject}
}
\description{
Compute the \code{tbl_sql} records of a Ramses object in the 
backend database (see \code{dplyr::\link[dplyr]{compute}()}). This function 
creates a database index of \code{Encounter} and \code{TherapyEpisode} 
longitudinal tables, which can speed up operations such as those performed
by \code{clinical_feature_x()} functions.
}
\seealso{
\code{\link[dplyr]{compute}()}
}
