% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{connect_local_database}
\alias{connect_local_database}
\title{Create or connect to a local DuckDB database}
\usage{
connect_local_database(file, timezone = Sys.timezone())
}
\arguments{
\item{file}{A file path to an existing or new database file with a
".duckdb" extension.}

\item{timezone}{A string for the time zone in which to return data to 
R from the database. By default, it is set to \code{\link{Sys.timezone}()}.}
}
\value{
A database connection object of class \code{duckdb_connection}.
}
\description{
Create a local database in memory or on disk using \code{\link[duckdb]{duckdb}()}. 
This is the ideal method to experiment on a small scale or for testing purposes. \code{duckdb} 
is a relational database similar to SQLite, with full support for date and datetime data.
}
\details{
This function creates a database on disk at the desired path. 
The database and its content will persist after it is disconnected.
}
\section{Warning}{
     
This method does not provide any encryption or password protection. You should only use this
method with mock data unless you operate within a secure data enclave.
}

\examples{
    # Create database and load data
    con <- connect_local_database("ramses-db.duckdb")
    
    dplyr::copy_to(dest = con, df = reference_aware, name = "reference_aware", 
                   overwrite = FALSE, temporary = FALSE)      
    
    # Close connection to database
    DBI::dbDisconnect(con, shutdown=TRUE)
    
    # Connect to the database again and show data
    con <- connect_local_database("ramses-db.duckdb")
    dplyr::tbl(con, "reference_aware")
    DBI::dbDisconnect(con, shutdown=TRUE)
    file.remove("ramses-db.duckdb")
}
\seealso{
The \code{duckdb} website provides excellent guidance on how to
connect to databases: 
\url{https://duckdb.org/docs/api/r}
}
