% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{create_therapy_episodes}
\alias{create_therapy_episodes}
\title{Re-create therapy episodes and therapy combinations}
\usage{
create_therapy_episodes(
  conn,
  transitive_closure_controls = transitive_closure_control(),
  silent = FALSE
)
}
\arguments{
\item{conn}{a connection to a database containing a
\code{drug_prescriptions} table previously created with 
\code{\link{load_medications}()})}

\item{transitive_closure_controls}{parameters controlling (see 
\code{\link{transitive_closure_control}()})}

\item{silent}{if \code{TRUE}, the progress bar will be hidden. The default is 
\code{FALSE}.}
}
\description{
This function should be used to recreate therapy episodes
and therapy combinations, for example in order to try new 
\code{transitive_closure_controls} parameters.
}
\section{Overview}{


Electronic prescribing and administration systems do not usually 
define:
\itemize{
  \item combination therapy: prescribing two or more antimicrobials  
  concurrently for the same indication
  \item therapy episodes: the set of antimicrobial prescriptions  
  administered consecutively or concurrently and corresponding to 
  a single indication and intent. 
}

When loading medication records, \code{\link{load_medications}}() first examines all 
prescriptions authored in a given hospital admission to ascertain such 
links between prescriptions. This involves looking at patterns of 
overlap and succession of prescriptions, and using a graph theory method 
known as 'transitive closure'.

Ramses links prescriptions together if:
\itemize{
  \item{prescriptions share the same \code{antiinfective_type}: 
  antibacterials are linked with antibacterials, antifungals with 
  antifungals, etc.}
  \item{\code{prescription_status} is *not* \code{'cancelled'},  
  \code{'draft'}, \code{'entered-in-error'}, or \code{'unknown'}.}
}
}

\section{Operations}{


This function performs the following operations:
\enumerate{
   \item verify that a valid \code{drug_prescriptions} table exists
   \item delete any existing \code{drug_prescription_edges} and
   \code{drug_therapy_episodes} table
   \item recreate the \code{drug_prescription_edges} table
   \item overwrite \code{therapy_id} and \code{combination_id}
   fields in \code{drug_prescriptions} table
   \item recreate the \code{drug_therapy_episodes} table
}
}

\seealso{
For more details please consult 
\code{vignette("therapy-episodes", package = "Ramses")}
}
