% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_records.R
\docType{data}
\name{drug_administrations}
\alias{drug_administrations}
\title{Simulated antimicrobial drug administrations}
\format{
A data frame of records of drug administrations.
}
\source{
Simulated data
}
\usage{
drug_administrations
}
\description{
A set of simulated records of administrations of antibiotics 
during inpatient care episodes, simulated based on drugs commonly 
prescribed in hospitals.
}
\keyword{datasets}
