% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_records.R
\docType{data}
\name{drug_prescriptions}
\alias{drug_prescriptions}
\title{Simulated antimicrobial prescriptions}
\format{A data frame of records of drug prescriptions.}
\source{
Simulated data
}
\usage{
drug_prescriptions
}
\description{
A set of drug prescriptions records simulated 
based on commonly prescribed drugs for clinical diagnoses 
recorded in inpatient care episode records.
}
\keyword{datasets}
