% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{import_icd}
\alias{import_icd}
\title{Import the International Classification of Diseases from an archive}
\usage{
import_icd(archive, version)
}
\arguments{
\item{archive}{path to the ZIP archive containing an ICD-10 release 
(5th Edition is recommended)}

\item{version}{a character string labelling the edition 
(e.g. "GB 5th Edition")}
}
\value{
A data frame ready for importing into the database with the 
following variables:
\describe{
    \item{icd_code}{character vector of ICD-10 codes without punctuation. 
    May include \code{"X"} placeholder characters depending on the edition.}
    \item{icd_display}{character vector of ICD-10 codes with a dot after the first character,
    for display in user interfaces.}
    \item{icd_description}{charactor vector of ICD-10 code descriptions}
    \item{category_code}{character vector of three-character ICD-10 codes}
    \item{category_description}{charactor vector three-character ICD-10 code descriptions}
    \item{edition}{edition label provided in parameter \code{version}}
    \item{\code{...}}{any other variables present in the \code{archive} source}
}
}
\description{
The World Health Organisation's (WHO) International 
Statistical Classification of Diseases and Related Health 
Problems Tenth Revision is available under restrictive licence terms.

The UK edition of ICD-10 is available from the NHS TRUD website
\url{https://isd.digital.nhs.uk/trud3/user/guest/group/0/pack/28} 
at no cost to UK users subject to a licence agreement.
 
The US edition of ICD-10-CM is available from 
\url{https://www.cdc.gov/nchs/icd/icd10cm.htm}
}
\examples{
\dontrun{
import_icd("icd_df_10.5.0_20151102000001.zip", "GB 5th Edition")
}
}
