% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_records.R
\docType{data}
\name{inpatient_diagnoses}
\alias{inpatient_diagnoses}
\title{Simulated inpatient clinical diagnoses}
\format{
A data frame of records of clinical diagnoses.
}
\source{
Simulated data
}
\usage{
inpatient_diagnoses
}
\description{
Set of simulated clinical diagnoses of infection for every
row in \code{\link[Ramses]{inpatient_episodes}}, patient, encounter, and 
episode numbers, along with up to 5 clinical diagnoses per admission
coded according to the International 
Classification of Diseases, Tenth Revision (ICD-10). 

This data was simulated to reproduce a set of combinations of infection 
diagnoses in an English NHS hospital. For each episode, a maximum of 5
diagnoses are simulated. Non-infection diagnoses are not simulated, an
NA value is recorded instead. 
 
This data was simulated to reproduce relevant characteristics of English 
NHS Admitted Patient Care datasets. More information can be found on the 
\href{https://datadictionary.nhs.uk/data_sets/cds_v6-2/cds_v6-2_type_130_-_admitted_patient_care_-_finished_general_episode_cds.html#dataset_cds_v6-2_type_130_-_admitted_patient_care_-_finished_general_episode_cds}{NHS data dictionary website}.
}
\keyword{datasets}
