% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_records.R
\docType{data}
\name{inpatient_episodes}
\alias{inpatient_episodes}
\title{Simulated episodes of inpatient care}
\format{
A data frame of records of finished consultant episodes.
}
\source{
Simulated data
}
\usage{
inpatient_episodes
}
\description{
This simulated activity data consists of records of finished 
consultant episodes for 200 patients, detailing all dates and times of 
admission, discharge, activity dates, and ICD-10 clinical diagnoses codes 
(for infections only).

This data was simulated to reproduce relevant characteristics of English 
NHS Admitted Patient Care datasets. More information can be found on the 
\href{https://www.datadictionary.nhs.uk/data_dictionary/messages/cds_v6-2/data_sets/cds_v6-2_type_130_-_admitted_patient_care_-_finished_general_episode_cds_fr.asp?shownav=1}{NHS data dictionary website}.
}
\keyword{datasets}
