% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_records.R
\docType{data}
\name{inpatient_microbiology}
\alias{inpatient_microbiology}
\title{Simulated microbial culture and susceptibility results}
\format{
A data frame
}
\source{
Simulated data
}
\usage{
inpatient_microbiology
}
\description{
A set of simulated records of microbial samples sent 
for culture and susceptibility testing for one patient included
in \code{\link[Ramses]{inpatient_episodes}}.
}
\seealso{
\link[Ramses]{validate_microbiology}()
}
\keyword{datasets}
