% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_records.R
\docType{data}
\name{inpatient_wards}
\alias{inpatient_wards}
\title{Simulated hospital ward movements}
\format{
A data frame of records of ward movements
}
\source{
Simulated data
}
\usage{
inpatient_wards
}
\description{
Set of simulated records of ward movements during 
hospitalisation encounters included in \code{\link[Ramses]{inpatient_episodes}}.

This data was simulated to reproduce relevant characteristics of English 
NHS Admitted Patient Care datasets. More information can be found on the 
\href{https://datadictionary.nhs.uk/data_sets/cds_v6-2/cds_v6-2_type_130_-_admitted_patient_care_-_finished_general_episode_cds.html#dataset_cds_v6-2_type_130_-_admitted_patient_care_-_finished_general_episode_cds}{NHS data dictionary website}.
}
\keyword{datasets}
