% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{load_inpatient_episodes}
\alias{load_inpatient_episodes}
\title{Load episodes of care records into the warehouse}
\usage{
load_inpatient_episodes(conn, episodes_data, overwrite = TRUE)
}
\arguments{
\item{conn}{a database connection}

\item{episodes_data}{a data frame validated with 
\code{\link{validate_inpatient_episodes}()}}

\item{overwrite}{if `TRUE` (the default), will overwrite any existing
`inpatient_episodes` database table}
}
\value{
`TRUE` if the function ran successfully, otherwise object of class
"try-error" containing error messages trigger during warehouse data loading.
}
\description{
Load episodes of care records into the warehouse
}
