% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{load_medications}
\alias{load_medications}
\title{Load drug prescription records into the warehouse}
\usage{
load_medications(
  conn,
  prescriptions,
  administrations = NULL,
  overwrite = FALSE,
  transitive_closure_controls = transitive_closure_control(),
  silent = TRUE
)
}
\arguments{
\item{conn}{a database connection}

\item{prescriptions}{a data frame of prescriptions passing the 
\code{\link{validate_prescriptions}()} function. Note: if variables 
`therapy_id` or `combination_id` are provided, they will be preserved
as they are instead of being populated by `Ramses`}

\item{administrations}{a data frame of drug administrations passing the 
\code{\link{validate_administrations}()} function}

\item{overwrite}{if `TRUE` (the default), will overwrite an existing 
diagnosis table}

\item{transitive_closure_controls}{parameters controlling (see 
\code{\link{transitive_closure_control}()})}

\item{silent}{a boolean indicating whether the function should be
executed without progress bar. Default is `TRUE`.}
}
\value{
`TRUE` if the function ran successfully, otherwise object of class
"try-error" containing error messages trigger during warehouse data loading.
}
\description{
Load drug prescription records into the warehouse
}
