% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{Patient-class}
\alias{Patient-class}
\alias{Patient}
\title{An S4 class to represent patients}
\usage{
Patient(conn, id)
}
\arguments{
\item{conn}{a database connection}

\item{id}{a patient identifier}
}
\description{
An S4 class to represent patients
}
\section{Slots}{

\describe{
\item{\code{id}}{a patient identifier}

\item{\code{conn}}{a database connection}

\item{\code{record}}{a \code{tbl_sql} for the corresponding database record}
}}

