% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_records.R
\docType{data}
\name{patients}
\alias{patients}
\title{Simulated patients}
\format{
A data frame of patient demographics.
}
\source{
Simulated data
}
\usage{
patients
}
\description{
Set of simulated demographics records of 202 patients 
included in \code{\link[Ramses]{inpatient_episodes}}.
}
\keyword{datasets}
