% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\docType{data}
\name{reference_loinc}
\alias{reference_loinc}
\title{LOINC codes for clinical investigations and observations}
\format{
A data frame.
}
\source{
LOINC Table File Version 2.70. Released 2021-06-10. \url{http://loinc.org}
}
\usage{
reference_loinc
}
\description{
Look-up table providing a subset of useful clinical investigation
codes from the LOINC terminology. It can be used to map clinical investigations
and observations that are relevant to modelling antibiotic prescribing decisions.
}
\details{
This material contains content from LOINC (\url{http://loinc.org}). 
LOINC is copyright © 1995-2021, Regenstrief Institute, Inc. and the 
Logical Observation Identifiers Names and Codes (LOINC) Committee and 
is available at no cost under the license at \url{http://loinc.org/license}.
LOINC® is a registered United States trademark of Regenstrief Institute, Inc.
}
\keyword{datasets}
