% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline.R
\name{therapy_timeline}
\alias{therapy_timeline}
\alias{therapy_timeline,Patient-method}
\alias{therapy_timeline,Encounter-method}
\alias{therapy_timeline,TherapyEpisode-method}
\title{Display an HTML antimicrobial therapy timeline}
\usage{
therapy_timeline(x, ...)

\S4method{therapy_timeline}{Patient}(
  x,
  date1 = NULL,
  date2 = NULL,
  load_timevis_dependencies = FALSE
)

\S4method{therapy_timeline}{Encounter}(x, load_timevis_dependencies = FALSE)

\S4method{therapy_timeline}{TherapyEpisode}(x, load_timevis_dependencies = FALSE)
}
\arguments{
\item{x}{an object of class \code{Patient}, \code{Encounter}, 
or \code{TherapyEpisode}}

\item{...}{other arguments (not in use)}

\item{date1}{(optional, only used for \code{x} of class \code{Patient}) 
a \code{Date} or \code{POSIXct} window minimum to 
focus the timeline on by default}

\item{date2}{(optional, only used for \code{x} of class \code{Patient}) 
a \code{Date} or \code{POSIXct} window maximum to 
focus the timeline on by default}

\item{load_timevis_dependencies}{if \code{TRUE}, jQuery and Bootstrap will
be loaded (default is \code{FALSE}). See \link[timevis]{timevis} for detail.}
}
\value{
a timevis object
}
\description{
Display an HTML antimicrobial therapy timeline
}
