% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{TherapyEpisode-class}
\alias{TherapyEpisode-class}
\alias{TherapyEpisode}
\alias{TherapyEpisode.DBIConnection}
\alias{TherapyEpisode.RamsesObject}
\title{An S4 class to represent therapy episodes}
\usage{
TherapyEpisode(...)

TherapyEpisode.DBIConnection(conn, id)

TherapyEpisode.RamsesObject(object)
}
\arguments{
\item{...}{generic signature}

\item{conn}{a database connection}

\item{id}{a vector of one or several therapy episode identifiers (from
database field \code{drug_therapy_episodes.therapy_id} as generated by 
\code{\link{load_medications}()} or
\code{\link{create_therapy_episodes}()})}

\item{object}{an object of class \code{MedicationRequest} or 
\code{Prescription}}
}
\description{
An S4 class to represent therapy episodes
}
\section{Slots}{

\describe{
\item{\code{id}}{a therapy episode identifier}

\item{\code{conn}}{a database connection}

\item{\code{record}}{a \code{tbl_sql} for the corresponding database record}

\item{\code{longitudinal_table}}{a \code{tbl_sql} for the longitudinal therapy table}
}}

