% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{TherapyEpisode-class}
\alias{TherapyEpisode-class}
\alias{TherapyEpisode}
\alias{TherapyEpisode.DBIConnection}
\alias{TherapyEpisode.RamsesObject}
\title{An S4 class to represent therapy episodes}
\usage{
TherapyEpisode(...)

TherapyEpisode.DBIConnection(conn, id, extend_table_start = NULL)

TherapyEpisode.RamsesObject(object, extend_table_start = NULL)
}
\arguments{
\item{...}{generic signature}

\item{conn}{a database connection}

\item{id}{a vector of one or several therapy episode identifiers (from
database field \code{drug_therapy_episodes.therapy_id} as generated by 
\code{\link{load_medications}()} or
\code{\link{create_therapy_episodes}()})}

\item{extend_table_start}{optional integer to specify an earlier start
(in hours) in the longitudinal table of the object. For example, a value of 
6 means the longitudinal table will begin 6 hours prior to the start of 
antimicrobial therapy. The value must be a positive number. Decimal numbers 
will be rounded up to the nearest integer. The default is \code{NULL}.}

\item{object}{an object of class \code{MedicationRequest} or 
\code{Prescription}}
}
\description{
An S4 class to represent therapy episodes
}
\section{Slots}{

\describe{
\item{\code{id}}{a therapy episode identifier}

\item{\code{conn}}{a database connection}

\item{\code{record}}{a \code{tbl_sql} for the corresponding database record}

\item{\code{longitudinal_table}}{a \code{tbl_sql} for the longitudinal therapy table}
}}

