% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{transitive_closure_control}
\alias{transitive_closure_control}
\title{Set parameters for building combination therapy identifiers and
therapy episode identifiers}
\usage{
transitive_closure_control(
  max_continuation_gap = 36,
  max_combination_authoring_gap = 6,
  max_combination_start_gap = 24
)
}
\arguments{
\item{max_continuation_gap}{a positive integer setting the maximum number of hours
tolerated between the end of a prescription and the start of a subsequent 
prescription in the same therapy episode. The default is 36 hours.}

\item{max_combination_authoring_gap}{a positive integer setting the maximum number
of hours tolerated between the time of authoring of two prescriptions 
administered as combination therapy. The default is 6 hours.}

\item{max_combination_start_gap}{a positive integer setting the maximum number of 
hours tolerated between the start of administration of two prescriptions
administred as combination therapy. The default is 24 hours.}
}
\value{
A list of three positive integer variables.
}
\description{
Set parameters controlling how prescriptions are 
linked into combinations and therapy episodes
}
\examples{
# default parameters
transitive_closure_control(
  max_continuation_gap = 36,
  max_combination_authoring_gap = 6,
  max_combination_start_gap = 24)
}
\seealso{
\code{\link[Ramses]{create_therapy_episodes}()}. For more details please consult 
\code{vignette("therapy-episodes", package = "Ramses")}
}
