% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_data.R
\name{validate_administrations}
\alias{validate_administrations}
\title{Validate medication administration records}
\usage{
validate_administrations(data)
}
\arguments{
\item{data}{a data frame containing one row per drug administration}
}
\value{
NULL if the `data` passes the validation. The function will trigger errors
for mandatory requirements and warnings for optional requirements.
}
\description{
This function performs a series of checks for mandatory and
optional requirements on drug administrations data.
}
\section{Mandatory fields}{

The following fields are required in order to pass the validation:
\describe{
     \item{\code{patient_id}}{a patient identifier with no missing value}
     \item{\code{prescription_id}}{a prescription identifier with no missing value}
     \item{\code{administration_id}}{an administration identifier with no missing value}
     \item{\code{administration_text}}{a character string summarising the drug to administer 
     (to be displayed in user interfaces, eg: \code{'Amoxicillin Oral 500mg'})}
     \item{\code{drug_id}}{identifier of the drug (from a dictionary such as SNOMED-CT or
      from \code{\link{AMR}{as.ab}()})}
     \item{\code{drug_name}}{preferred name of the drug in the drug dictionary}
     \item{\code{drug_display_name}}{drug name to display in reports and user interfaces
     (can be the same as \code{drug_name})}
     \item{\code{ATC_code}}{the ATC code, see \code{\link{AMR}{ab_atc}()}}
     \item{\code{ATC_group}}{the ATC group, see \code{\link{AMR}{ab_group1}()}}
     \item{\code{ATC_route}}{route of administration as defined in the ATC ("O" = oral; 
     "P" = parenteral; "R" = rectal; "V" = vaginal)}
     \item{\code{dose}}{a numeric vector of dosage quantities}
     \item{\code{unit}}{a character vector of dosage units}
     \item{\code{route}}{the route of administration value natively assigned by system}
     \item{\code{administration_date}}{timestamp of the drug administration}
     \item{\code{administration_status}}{one value from the following 
     \href{https://hl7.org/fhir/R4/valueset-medicationrequest-status.html}{FHIR R4}
     reference set:
     \itemize{ 
        \item \code{`in-progess`} the administration has started but has not yet completed.
        \item \code{`not-done`} the administration was terminated prior to any impact on 
        the subject (though preparatory actions may have been taken). 
        \item \code{`on-hold`} actions implied by the administration have been 
        temporarily halted, but are expected to continue later.
        \item \code{`completed`} all actions that are implied by the administration 
        have occurred.
        \item \code{`entered-in-error`} the administration was entered in error and 
        therefore nullified. 
        \item \code{`stopped`} actions implied by the administration have been permanently 
        halted, before all of them occurred.
        \item \code{`unknown`} the authoring/source system does not know which of the 
        status values currently applies for this request. \emph{Note:} This concept 
        is not to be used for 'other' - one of the listed statuses is presumed 
        to apply, but the authoring/source system does not know which.         
}}}
}

\section{Optional fields}{

\describe{
     \item{\code{DDD}}{the number of defined daily doses (as administered),
      see \code{\link{compute_DDDs}()}}
     \item{\code{...}}{any other field, as desired, can be loaded into the database.}}
}

