% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_data.R
\name{validate_inpatient_diagnoses}
\alias{validate_inpatient_diagnoses}
\title{Validate inpatient diagnosis records}
\usage{
validate_inpatient_diagnoses(diagnoses_data, diagnoses_lookup)
}
\arguments{
\item{diagnoses_data}{a data frame containing clinical diagnoses, with, 
at minimum, variables \code{patient_id}, \code{encounter_id}, 
\code{episode_number}, \code{icd_code}, \code{diagnosis_position}}

\item{diagnoses_lookup}{a data frame containing an ICD-10 reference look up 
table with, at minimum, variables \code{icd_description}, \code{icd_display}, 
\code{category_code}, \code{category_description}}
}
\value{
A logical value indicating success
}
\description{
Validate constraints on diagnosis records, namely that the 
minimum variables are present, and that all \code{icd_code} values can be 
looked up in an ICD-10 reference table
}
\section{Diagnoses set mandatory variables}{

\describe{
  \item{\code{patient_id}}{a patient identifier with no missing value}
  \item{\code{encounter_id}}{a hospital encounter identifier with no missing value}
  \item{\code{episode_number}}{a strictly positive integer indicating the
  number of the episode within an admission. Must not be missing.}
  \item{\code{icd_code}}{a code corresponding to the International Classification of Diseases
  without a "." separator}
  \item{\code{diagnosis_position}}{an integer describing the diagnosis position
  on the discharge summary (1 = primary cause of admission}
}
}

\section{Diagnoses lookup mandatory variables}{


\describe{
  \item{\code{icd_code}}{character diagnosis codes corresponding to the 
  International Classification of Diseases formatted without a "." separator (eg "A0101")}
  \item{\code{icd_display}}{laid-out diagnosis codes for display (eg: "A01.01")}
  \item{\code{icd_description}}{full text descriptions of the diagnoses}
  \item{\code{category_code}}{three-character heading codes (eg "A01")}
  \item{\code{category_description}}{full text descriptions of three-character heading codes}
}

*Note*: \code{\link[Ramses]{import_icd}()} can produce this lookup data frame from
a standard ICD release archive file.
}

\examples{
data_icd <- dplyr::filter(Ramses::inpatient_diagnoses, !is.na(icd_code))
lookup_icd <- dplyr::distinct(data_icd, icd_code)
lookup_icd$icd_display <- lookup_icd$icd_code
lookup_icd$icd_description <- "ICD-10 code label text"
lookup_icd$category_code <- substr(lookup_icd$icd_code, 0, 3)
lookup_icd$category_description <- "ICD-10 category label text"
validate_inpatient_diagnoses(data_icd, lookup_icd)
}
