% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_data.R
\name{validate_inpatient_diagnoses}
\alias{validate_inpatient_diagnoses}
\title{Validate inpatient diagnosis records}
\usage{
validate_inpatient_diagnoses(diagnoses_data, diagnoses_lookup)
}
\arguments{
\item{diagnoses_data}{a data frame containing clinical diagnoses, with, 
at minimum, variables \code{patient_id}, \code{spell_id}, 
\code{episode_number}, \code{icd_code}, \code{diagnosis_position}}

\item{diagnoses_lookup}{a data frame containing an ICD-10 reference look up 
table with, at minimum, variables `icd_description`, `icd_display`, 
`category_code`, `category_description`}
}
\value{
A logical value indicating success
}
\description{
Validate constraints on diagnosis records, namely that the 
minimum variables are present, and that all \code{icd_code} values can be 
looked up in an ICD-10 reference table
}
\examples{
data_icd <- dplyr::filter(Ramses::inpatient_diagnoses, !is.na(icd_code))
lookup_icd <- dplyr::distinct(data_icd, icd_code)
lookup_icd$icd_display <- lookup_icd$icd_code
lookup_icd$icd_description <- "ICD-10 code label text"
lookup_icd$category_code <- substr(lookup_icd$icd_code, 0, 3)
lookup_icd$category_description <- "ICD-10 category label text"
validate_inpatient_diagnoses(data_icd, lookup_icd)
}
