% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_data.R
\name{validate_inpatient_episode_dates}
\alias{validate_inpatient_episode_dates}
\title{Validate the consistency of inpatient episode dates}
\usage{
validate_inpatient_episode_dates(data, type = "episodes")
}
\arguments{
\item{data}{a data frame object}

\item{type}{a string indicating the type of dates to validate:
either \code{"wards"} for ward stays or \code{"episodes"} for inpatient episodes.}
}
\value{
A logical value indicating success
}
\description{
Validate the consistency of inpatient episode dates
}
