% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_data.R
\name{validate_microbiology}
\alias{validate_microbiology}
\title{Validate microbial isolates & susceptibility records}
\usage{
validate_microbiology(specimens, isolates, susceptibilities)
}
\arguments{
\item{specimens}{a data frame with one row per specimen sent
to laboratory (see details)}

\item{isolates}{a data frame with one row per microorganism
isolated from the laboratory specimen (see details)}

\item{susceptibilities}{a data frame with one row per susceptibility
(see details)}
}
\value{
TRUE if the validation is passed
}
\description{
Validate microbial isolates & susceptibility records
}
\section{Specimens data frame}{

\strong{The following fields are mandatory:}
\describe{
   \item{\code{specimen_id}}{a unique specimen identifier with no missing value}
   \item{\code{patient_id}}{a patient identifier with no missing value}
   \item{\code{status}}{one value from the following 
   \href{https://www.hl7.org/fhir/valueset-specimen-status.html}{FHIR R4}
     reference set: \itemize{
       \item \code{"available"}: The physical specimen is present and 
       in good condition.
       \item \code{"unavailable"}: There is no physical specimen because it is either 
       lost, destroyed or consumed.
       \item \code{"unsatisfactory"}: The specimen cannot be used because of a
        quality issue such as a broken container, contamination, or too old.	
       \item \code{"entered-in-error"}: The specimen was entered in error and 
       therefore nullified.
     }}
   \item{\code{specimen_datetime}}{datetime when specimen was sampled or 
   received for processing.}
   \item{\code{specimen_type_code}}{character vector of descendants of 
   the SNOMED CT concept \code{123038009 | Specimen (specimen) |}. Admissible
   values are listed in \link[Ramses]{reference_specimen_type}}
   \item{\code{specimen_type_name}}{character vector of the SNOMED CT
   preferred terms for \code{specimen_type_code}}
   \item{\code{specimen_type_display}}{character vector of custom specimen
   types for display in user interfaces}
}

\strong{The following fields are optional:}
\describe{
   \item{\code{spell_id}}{a hospital spell identifier (if the specimen was 
   sampled during admission)}
   \item{\code{test_display}}{free text description of test requested for 
   display in user interfaces. For instance: "Mycobacteria culture" or 
   "Microbial culture, anaerobic, initial isolation". Coded concepts
   should be stored in other custom columns.}
   \item{\code{reason_display}}{free text reason for a procedure for display 
   in user interfaces. For instance, "Suspected urinary tract infection", or
   "Surgical site microbiological sample". Coded concepts should be stored in other 
   custom columns.}
}
}

\section{Isolates data frame}{


\strong{The following fields are mandatory:}
\describe{
   \item{\code{isolate_id}}{a unique isolated organism identifier with no missing value}
   \item{\code{specimen_id}}{a specimen identifier with no missing value}
   \item{\code{patient_id}}{a patient identifier with no missing value}
   \item{\code{organism_code}}{a character vector containing either: \itemize{
      \item a microorganism code validated using \code{\link[AMR]{as.mo}()}
      \item \code{NA_character_} if no microorganism was isolated, for instance
      due to no growth or mixed heavy growth
   }}
   \item{\code{organism_name}}{a microorganism name provided by 
   \code{\link[AMR]{mo_name}()}, or \code{NA} if no microorganism was isolated}
   \item{\code{organism_display_name}}{microorganism name as labelled by the
   laboratory, for display in user interfaces. No growth/mixed heavy growth should
   be referenced here}
   \item{\code{isolation_datetime}}{datetime when the organism was first isolated 
   (or reported) by the laboratory}
 }

\strong{The following field is optional:}
\describe{
   \item{\code{mdr_classification}}{character vector of classifications produced by 
   \code{\link[AMR]{mdro}()}. Admissible values are: \itemize{
      \item \code{NA_character_} when susceptibilities are not available/conclusive
      \item \code{"Negative"} for isolate presenting no wide resistance phenotype
      \item other character codes dependent on the \code{guideline} parameter 
      provided to \code{\link[AMR]{mdro}()}
   }}
}
}

\section{Susceptibilities data frame}{


\strong{The following fields are mandatory:}
\describe{
  \item{\code{isolate_id}}{an isolated organism identifier with no missing value}
  \item{\code{specimen_id}}{a specimen identifier with no missing value}
  \item{\code{patient_id}}{a patient identifier with no missing value}
  \item{\code{organism_code}}{a microorganism code validated using 
  \code{\link[AMR]{as.mo}()}, with no missing values}
  \item{\code{organism_name}}{a microorganism name provided by 
  \code{\link[AMR]{mo_name}()}, with no missing values}
  \item{\code{organism_display_name}}{microorganism name as labelled by the
  laboratory, for display in user interfaces, with no missing values}
  \item{\code{agent_code}}{code of the antimicrobial tested as provided by 
  \code{\link[AMR]{as.ab}()}}
  \item{\code{agent_name}}{name of the antimicrobial tested as provided by 
  \code{\link[AMR]{ab_name}()}}
  \item{\code{agent_display_name}}{name of the antimicrobial tested, 
  with no missing values, for  display in reports and user interfaces
  (can be the same as \code{drug_name})}
  \item{\code{rsi_code}}{\code{"R"} (resistant), \code{"S"} (susceptible), 
   or \code{"I"} (intermediate exposure), as determined by the laboratory or by
  \code{\link[AMR]{as.rsi}()} on the basis of minimum inhibitory concentrations
  or disk diffusion diameters}
}
}

